requirements.txt
pandas==2.2.3
openpyxl==3.1.5
selenium==4.26.0
requests==2.32.3
backoff==2.2.1
beautifulsoup4==4.12.3
google-auth-oauthlib==1.2.1
google-api-python-client==2.149.0
oauth2client==4.1.3
pylint==3.3.1
Flask==3.0.3
gunicorn==23.0.0
PyPDF2==3.0.1
google-analytics-data==0.18.14
jsonlines==4.0.0
joblib==1.4.2
python-dotenv==1.0.1
playwright==1.48.0